refactor(S-3.01): extract handle_login to auth/login.rs
Move resolve_oauth_scopes, login_token, login_oauth, LoginArgs,
handle_login, and prepare_login_target out of mod.rs into login.rs.

login.rs imports credential constants directly from super::keychain
rather than via mod.rs re-exports. prepare_login_target and
resolve_oauth_scopes re-exported under #[cfg(test)] for unit tests.
mod.rs retains auth_json_response, AuthFlow, chosen_flow_for_profile
(used across multiple shards via super::) and adds #[cfg(test)] imports
for JrError, auth, OAuthAppSource needed by the test module.
612 unit tests pass.
